+++ example/lib/screens/home/bloc/home_communication.dart
@@ -1,7 +1,7 @@
 import 'package:bsev/bsev.dart';
 import 'package:bsev_demo/repository/pokemon/model/pokemon.dart';
 
-class HomeCommunication extends CommunicationBase {
+class HomeCommunication extends Communication {
   final pokemonList = BehaviorSubjectCreate<List<Pokemon>>();
   final showProgress = BehaviorSubjectCreate<bool>(initValue: false);