+++ lib/src/flavors.dart
@@ -1,0 +1,18 @@
+enum Flavor { DEBUG, HML, PROD }
+
+class Flavors {
+  static final Flavors _singleton = Flavors._internal();
+  static Flavor _flavor;
+
+  factory Flavors() {
+    return _singleton;
+  }
+
+  Flavors._internal();
+
+  static void configure(Flavor flavor) {
+    _flavor = flavor;
+  }
+
+  Flavor getFlavor() => _flavor;
+}